protocolVersion and move datatype in firmwareconstants
