+++ server/core/src/main/java/dev/slimevr/tracking/trackers/udp/TrackersUDPServer.kt
@@ -79,7 +79,7 @@ class TrackersUDPServer(private val port: Int, name: String, private val tracker
 				ipAddress = addr
 				name = handshake.macString?.let { "udp://$it" }
 				descriptiveName = "udp:/$addr"
-				firmwareBuild = handshake.firmwareBuild
+				protocolVersion = handshake.protocolVersion
 				firmwareVersion = handshake.firmware
 				connectionsByAddress[address] = this
 
@@ -90,7 +90,7 @@ class TrackersUDPServer(private val port: Int, name: String, private val tracker
 						[TrackerServer] Tracker $i handed over to address $socketAddr.
 						Board type: ${handshake.boardType},
 						imu type: ${handshake.imuType},
-						firmware: ${handshake.firmware} ($firmwareBuild),
+						firmware: ${handshake.firmware} ($protocolVersion),
 						mac: ${handshake.macString},
 						name: $name
 						""".trimIndent(),
@@ -103,7 +103,7 @@ class TrackersUDPServer(private val port: Int, name: String, private val tracker
 				name = handshake.macString?.let { "udp://$it" }
 					?: "udp:/$addr"
 				descriptiveName = "udp:/$addr"
-				firmwareBuild = handshake.firmwareBuild
+				protocolVersion = handshake.protocolVersion
 				firmwareVersion = handshake.firmware
 				val i = connections.indexOf(this)
 				LogManager
@@ -112,7 +112,7 @@ class TrackersUDPServer(private val port: Int, name: String, private val tracker
 						[TrackerServer] Tracker $i reconnected from address $socketAddr.
 						Board type: ${handshake.boardType},
 						imu type: ${handshake.imuType},
-						firmware: ${handshake.firmware} ($firmwareBuild),
+						firmware: ${handshake.firmware} ($protocolVersion),
 						mac: ${handshake.macString},
 						name: $name
 						""".trimIndent(),
@@ -129,7 +129,7 @@ class TrackersUDPServer(private val port: Int, name: String, private val tracker
 				handshake.mcuType,
 			)
 			VRServer.instance.deviceManager.addDevice(connection)
-			connection.firmwareBuild = handshake.firmwareBuild
+			connection.protocolVersion = handshake.protocolVersion
 			connection.protocol = if (handshake.firmware?.isEmpty() == true) {
 				// Only old owoTrack doesn't report firmware and have different packet IDs with SlimeVR
 				NetworkProtocol.OWO_LEGACY
@@ -158,13 +158,13 @@ class TrackersUDPServer(private val port: Int, name: String, private val tracker
 						[TrackerServer] Tracker $i connected from address $socketAddr.
 						Board type: ${handshake.boardType},
 						imu type: ${handshake.imuType},
-						firmware: ${handshake.firmware} (${connection.firmwareBuild}),
+						firmware: ${handshake.firmware} (${connection.protocolVersion}),
 						mac: ${handshake.macString},
 						name: ${connection.name}
 						""".trimIndent(),
 					)
 			}
-			if (connection.protocol == NetworkProtocol.OWO_LEGACY || connection.firmwareBuild < 9) {
+			if (connection.protocol == NetworkProtocol.OWO_LEGACY || connection.protocolVersion < 9) {
 				// Set up new sensor for older firmware.
 				// Firmware after 7 should send sensor status packet and sensor
 				// will be created when it's received
